Mabruka/requirements.txt
Django
django-braces
django-filter
django-mptt
djangorestframework
drfdocs
gunicorn
Markdown
SwampDragon
Unipath
whitenoise
